Refactor StorageCluster creation into it's own module
