+++ .secrets.baseline
@@ -428,16 +428,6 @@
         "verified_result": null
       }
     ],
-    "ocs_ci/deployment/deployment.py": [
-      {
-        "hashed_secret": "f0e2d8610edefa0c02b673dcac7964b02ce3e890",
-        "is_secret": false,
-        "is_verified": false,
-        "line_number": 3180,
-        "type": "Basic Auth Credentials",
-        "verified_result": null
-      }
-    ],
     "ocs_ci/deployment/flexy.py": [
       {
         "hashed_secret": "bea1af2a15561a266eda3472673d24cb4f2020d4",
@@ -902,21 +892,13 @@
         "verified_result": null
       }
     ],
-    "ocs_ci/workloads/gosbench_workload.py": [
+    "ocs_ci/utility/storage_cluster_setup.py": [
       {
-        "hashed_secret": "a83d4bf9070ae6eb080b4cc7b2703e17a51172ae",
+        "hashed_secret": "f0e2d8610edefa0c02b673dcac7964b02ce3e890",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 1290,
-        "type": "Secret Keyword",
-        "verified_result": null
-      },
-      {
-        "hashed_secret": "2739bb260ce45a5ad560051592efe21fb19aca85",
-        "is_secret": false,
-        "is_verified": false,
-        "line_number": 1338,
-        "type": "Secret Keyword",
+        "line_number": 586,
+        "type": "Basic Auth Credentials",
         "verified_result": null
       }
     ],
